Move `analytics` to the `lib` folder
