+++ lib/analytics.ts
